+++ src/main/java/com/twilio/rest/flexapi/v1/InsightsQuestionnairesCategoryDeleter.java
@@ -14,7 +14,7 @@
 
 package com.twilio.rest.flexapi.v1;
 
-import com.twilio.base.Fetcher;
+import com.twilio.base.Deleter;
 import com.twilio.exception.ApiConnectionException;
 import com.twilio.exception.ApiException;
 import com.twilio.exception.RestException;
@@ -26,25 +26,27 @@ import com.twilio.rest.Domains;
 
 
 
-
-public class UserRolesFetcher extends Fetcher<UserRoles> {
+public class InsightsQuestionnairesCategoryDeleter extends Deleter<InsightsQuestionnairesCategory> {
+    private String pathCategoryId;
     private String token;
 
-    public UserRolesFetcher(){
+    public InsightsQuestionnairesCategoryDeleter(final String pathCategoryId){
+        this.pathCategoryId = pathCategoryId;
     }
 
-    public UserRolesFetcher setToken(final String token){
+    public InsightsQuestionnairesCategoryDeleter setToken(final String token){
         this.token = token;
         return this;
     }
 
     @Override
-    public UserRoles fetch(final TwilioRestClient client) {
-        String path = "/v1/Insights/UserRoles";
+    public boolean delete(final TwilioRestClient client) {
+        String path = "/v1/Insights/QM/Categories/{CategoryId}";
 
+        path = path.replace("{"+"CategoryId"+"}", this.pathCategoryId.toString());
 
         Request request = new Request(
-            HttpMethod.GET,
+            HttpMethod.DELETE,
             Domains.FLEXAPI.toString(),
             path
         );
@@ -52,7 +54,7 @@ public class UserRolesFetcher extends Fetcher<UserRoles> {
         Response response = client.request(request);
 
         if (response == null) {
-        throw new ApiConnectionException("UserRoles fetch failed: Unable to connect to server");
+            throw new ApiConnectionException("InsightsQuestionnairesCategory delete failed: Unable to connect to server");
         } else if (!TwilioRestClient.SUCCESS.test(response.getStatusCode())) {
             RestException restException = RestException.fromJson(response.getStream(), client.getObjectMapper());
             if (restException == null) {
@@ -60,8 +62,7 @@ public class UserRolesFetcher extends Fetcher<UserRoles> {
             }
             throw new ApiException(restException);
         }
-
-        return UserRoles.fromJson(response.getStream(), client.getObjectMapper());
+        return response.getStatusCode() == 204;
     }
     private void addHeaderParams(final Request request) {
         if (token != null) {